Remove AddRequestPolicyBefore/After. UserAgentPolicyFactory inversion.
