+++ client-runtime/src/test/java/com/microsoft/rest/v2/UserAgentTests.java
@@ -6,14 +6,13 @@
 
 package com.microsoft.rest.v2;
 
-import com.microsoft.rest.v2.http.HttpClient;
 import com.microsoft.rest.v2.http.HttpMethod;
 import com.microsoft.rest.v2.http.HttpPipeline;
 import com.microsoft.rest.v2.http.HttpRequest;
 import com.microsoft.rest.v2.http.HttpResponse;
 import com.microsoft.rest.v2.http.MockHttpClient;
 import com.microsoft.rest.v2.http.MockHttpResponse;
-import com.microsoft.rest.v2.policy.UserAgentPolicy;
+import com.microsoft.rest.v2.policy.UserAgentPolicyFactory;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -34,7 +33,7 @@ public class UserAgentTests {
                     return Single.<HttpResponse>just(new MockHttpResponse(200));
                 }
             },
-            new UserAgentPolicy.Factory("AutoRest-Java"));
+            new UserAgentPolicyFactory("AutoRest-Java"));
 
         HttpResponse response = pipeline.sendRequestAsync(new HttpRequest(
                 "defaultUserAgentTests",
@@ -54,7 +53,7 @@ public class UserAgentTests {
                     return Single.<HttpResponse>just(new MockHttpResponse(200));
                 }
             },
-            new UserAgentPolicy.Factory("Awesome"));
+            new UserAgentPolicyFactory("Awesome"));
 
         HttpResponse response = pipeline.sendRequestAsync(new HttpRequest("customUserAgentTests", HttpMethod.GET, new URL("http://localhost"))).blockingGet();
         Assert.assertEquals(200, response.statusCode());
