+++ client-runtime/src/test/java/com/microsoft/rest/v2/http/HttpPipelineBuilderTests.java
@@ -58,60 +58,6 @@ public class HttpPipelineBuilderTests {
     }
 
     @Test
-    public void withRequestPolicyBefore() {
-        final HttpPipelineBuilder builder = new HttpPipelineBuilder();
-
-        builder.withRequestPolicyBefore(RetryPolicy.Factory.class, new PortPolicy.Factory(80));
-        assertEquals(1, builder.requestPolicyFactories().size());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-
-        builder.withRequestPolicyBefore(RetryPolicy.Factory.class, new ProtocolPolicy.Factory("ftp"));
-        assertEquals(2, builder.requestPolicyFactories().size());
-        assertEquals(ProtocolPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
-
-        builder.withRequestPolicyBefore(ProtocolPolicy.Factory.class, new RetryPolicy.Factory());
-        assertEquals(3, builder.requestPolicyFactories().size());
-        assertEquals(ProtocolPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(RetryPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(2).getClass());
-
-        builder.withRequestPolicyBefore(PortPolicy.Factory.class, new LoggingPolicy.Factory(LoggingPolicy.LogLevel.BODY));
-        assertEquals(4, builder.requestPolicyFactories().size());
-        assertEquals(ProtocolPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(RetryPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(2).getClass());
-        assertEquals(LoggingPolicy.Factory.class, builder.requestPolicyFactories().get(3).getClass());
-    }
-
-    @Test
-    public void withRequestPolicyAfter() {
-        final HttpPipelineBuilder builder = new HttpPipelineBuilder();
-
-        builder.withRequestPolicyAfter(RetryPolicy.Factory.class, new PortPolicy.Factory(80));
-        assertEquals(1, builder.requestPolicyFactories().size());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-
-        builder.withRequestPolicyAfter(RetryPolicy.Factory.class, new ProtocolPolicy.Factory("ftp"));
-        assertEquals(2, builder.requestPolicyFactories().size());
-        assertEquals(ProtocolPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
-
-        builder.withRequestPolicyAfter(ProtocolPolicy.Factory.class, new RetryPolicy.Factory());
-        assertEquals(3, builder.requestPolicyFactories().size());
-        assertEquals(RetryPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(ProtocolPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(2).getClass());
-
-        builder.withRequestPolicyAfter(PortPolicy.Factory.class, new LoggingPolicy.Factory(LoggingPolicy.LogLevel.BODY));
-        assertEquals(4, builder.requestPolicyFactories().size());
-        assertEquals(RetryPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(ProtocolPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
-        assertEquals(LoggingPolicy.Factory.class, builder.requestPolicyFactories().get(2).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(3).getClass());
-    }
-
-    @Test
     public void withRequestPolicyArray() {
         final HttpPipelineBuilder builder = new HttpPipelineBuilder();
